refactor: move selectors to separate folders
