+++ Chat/src/components/Navigation/index.tsx
@@ -7,10 +7,11 @@ import RegistrationScreen from '../../screens/RegistrationScreen';
 import {useAppSelector} from '../../hooks/storeHooks';
 import ChatRoom from '../../screens/ChatRoom';
 import Profile from '../../screens/Profile';
+import {selectUser} from '../../store/slices/user/selectors';
 
 const Stack = createNativeStackNavigator<RootStackParamList>();
 function Navigation() {
-  const user = useAppSelector(state => state.user);
+  const user = useAppSelector(selectUser);
 
   return (
     <Stack.Navigator>
