+++ Chat/src/store/store.ts
@@ -1,8 +1,8 @@
 import {configureStore} from '@reduxjs/toolkit';
-import chatSlice from './slices/chatSlice';
-import messageSlice from './slices/messageSlice';
+import chatSlice from './slices/chat/chatSlice';
+import messageSlice from './slices/message/messageSlice';
 import socketSlice from './slices/socketSlice';
-import userSlice from './slices/userSlice';
+import userSlice from './slices/user/userSlice';
 
 export const store = configureStore({
   reducer: {
